requirements.txt
django
djangorestframework
markdown
django-filter
opencv-python
numpy
torch 
torchvision 
torchaudio
pillow
python-memcached
pandas
requests
PyYAML
tqdm
matplotlib
seaborn
thop